Añadida lista

+++ app/src/main/java/com/deneb/astroapps/wheateraac/data/db/entity/Day.kt
@@ -10,9 +10,9 @@ data class Day(
     @SerializedName("avgtemp_f")
     val avgtempF: Double,
     @SerializedName("avgvis_km")
-    val avgvisKm: Int,
+    val avgvisKm: Double,
     @SerializedName("avgvis_miles")
-    val avgvisMiles: Int,
+    val avgvisMiles: Double,
     @Embedded(prefix = "condition_")
     val condition: Condition,
     @SerializedName("maxtemp_c")
@@ -24,12 +24,12 @@ data class Day(
     @SerializedName("maxwind_mph")
     val maxwindMph: Double,
     @SerializedName("mintemp_c")
-    val mintempC: Int,
+    val mintempC: Double,
     @SerializedName("mintemp_f")
     val mintempF: Double,
     @SerializedName("totalprecip_in")
-    val totalprecipIn: Int,
+    val totalprecipIn: Double,
     @SerializedName("totalprecip_mm")
-    val totalprecipMm: Int,
-    val uv: Int
+    val totalprecipMm: Double,
+    val uv: Double
 )
